requirements.txt
configparser==3.5.0
coverage==4.5.1
crispy-forms-bulma==1.1.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==1.11.15
django-admin-sortable2==0.6.21
django-bulma==0.5.1
django-coverage==1.2.4
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-dotenv==1.4.2
django-extensions==2.1.2
django-filebrowser==3.10.1
django-filebrowser-no-grappelli==3.7.5
django-grappelli==2.11.1
django-js-asset==1.1.0
django-localflavor==2.1
django-mptt==0.9.1
django-nose==1.4.5
django-sitetree==1.11.0
-e git+http://github.com/waustin/django-text-blocks.git@4c4648cd0868150bcbb8615230574f1e15eb45b7#egg=django_text_blocks
djangorestframework==3.8.2
easy-thumbnails==2.5
enum34==1.1.6
factory-boy==2.11.1
Faker==0.9.0
flake8==3.5.0
ipaddress==1.0.22
mccabe==0.6.1
nose==1.3.7
pbr==1.10.0
pdfminer==20140328
pdfrw==0.4
pep8==1.7.1
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycodestyle==2.3.1
pyflakes==2.0.0
pygraphviz==1.3.1
PyPDF2==1.26.0
python-dateutil==2.7.3
pytz==2018.5
six==1.10.0
sqlparse==0.2.4
static3==0.7.0
stevedore==1.20.0
text-unidecode==1.2
typing==3.6.6
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
Werkzeug==0.14.1
